requirements.txt
Flask-JWT-Extended
Flask-RESTful
Flask-SQLAlchemy
marshmallow
flask-marshmallow
marshmallow-sqlalchemy
